+++ src/addons/api/b32.rs
