+++ src/addons/mod.rs
@@ -8,5 +8,5 @@ mod addresses;
 mod api;
 
 pub use addresses::mock::MockAddressGenerator;
-pub use api::bech32::MockApiBech32;
-pub use api::bech32m::MockApiBech32m;
+pub use api::b32::MockApiBech32;
+pub use api::b32m::MockApiBech32m;
